requirements.txt

matplotlib==3.9.2
numpy==2.1.3
statsmodel>=0.13.5
river>=0.18.0